+++ cmd/neofs-node/config.go
@@ -26,7 +26,6 @@ import (
 	blobovniczaconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/engine/shard/blobstor/blobovnicza"
 	fstreeconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/engine/shard/blobstor/fstree"
 	loggerconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/logger"
-	metricsconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/metrics"
 	nodeconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/node"
 	objectconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/object"
 	replicatorconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/replicator"
@@ -50,6 +49,7 @@ import (
 	"github.com/nspcc-dev/neofs-node/pkg/network"
 	"github.com/nspcc-dev/neofs-node/pkg/network/cache"
 	"github.com/nspcc-dev/neofs-node/pkg/services/control"
+	objectService "github.com/nspcc-dev/neofs-node/pkg/services/object"
 	getsvc "github.com/nspcc-dev/neofs-node/pkg/services/object/get"
 	"github.com/nspcc-dev/neofs-node/pkg/services/object_manager/tombstone"
 	tsourse "github.com/nspcc-dev/neofs-node/pkg/services/object_manager/tombstone/source"
@@ -308,7 +308,7 @@ type internals struct {
 
 	wg      *sync.WaitGroup
 	workers []worker
-	closers []func()
+	closers []closer
 
 	apiVersion   version.Version
 	healthStatus *atomic.Int32
@@ -363,12 +363,16 @@ type shared struct {
 	treeService *tree.Service
 
 	metricsCollector *metrics.NodeMetrics
+
+	metricsSvc *objectService.MetricCollector
 }
 
 // dynamicConfiguration stores parameters of the
 // components that supports runtime reconfigurations.
 type dynamicConfiguration struct {
-	logger *logger.Prm
+	logger  *logger.Prm
+	pprof   *httpComponent
+	metrics *httpComponent
 }
 
 type cfg struct {
@@ -607,10 +611,8 @@ func initCfg(appCfg *config.Config) *cfg {
 
 	user.IDFromKey(&c.ownerIDFromKey, key.PrivateKey.PublicKey)
 
-	if metricsconfig.Enabled(c.appCfg) {
-		c.metricsCollector = metrics.NewNodeMetrics()
-		netState.metrics = c.metricsCollector
-	}
+	c.metricsCollector = metrics.NewNodeMetrics()
+	netState.metrics = c.metricsCollector
 
 	c.onShutdown(c.clientCache.CloseAll)   // clean up connections
 	c.onShutdown(c.bgClientCache.CloseAll) // clean up connections
@@ -904,11 +906,9 @@ func (c *cfg) ObjectServiceLoad() float64 {
 	return float64(c.cfgObject.pool.putRemote.Running()) / float64(c.cfgObject.pool.putRemoteCapacity)
 }
 
-type dCfg struct {
-	name string
-	cfg  interface {
-		Reload() error
-	}
+type dCmp struct {
+	name       string
+	reloadFunc func() error
 }
 
 func (c *cfg) configWatcher(ctx context.Context) {
@@ -928,7 +928,7 @@ func (c *cfg) configWatcher(ctx context.Context) {
 
 			// all the components are expected to support
 			// Logger's dynamic reconfiguration approach
-			var components []dCfg
+			var components []dCmp
 
 			// Logger
 
@@ -938,7 +938,18 @@ func (c *cfg) configWatcher(ctx context.Context) {
 				continue
 			}
 
-			components = append(components, dCfg{name: "logger", cfg: logPrm})
+			components = append(components, dCmp{"logger", logPrm.Reload})
+			if cmp, updated := metricsComponent(c); updated {
+				if cmp.enabled {
+					cmp.preReload = enableMetricsSvc
+				} else {
+					cmp.preReload = disableMetricsSvc
+				}
+				components = append(components, dCmp{cmp.name, cmp.reload})
+			}
+			if cmp, updated := pprofComponent(c); updated {
+				components = append(components, dCmp{cmp.name, cmp.reload})
+			}
 
 			// Storage Engine
 
@@ -954,7 +965,7 @@ func (c *cfg) configWatcher(ctx context.Context) {
 			}
 
 			for _, component := range components {
-				err = component.cfg.Reload()
+				err = component.reloadFunc()
 				if err != nil {
 					c.log.Error("updated configuration applying",
 						zap.String("component", component.name),